# Default module.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Default module",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/users/{userId}": {
      "post": {
        "summary": "유저 정보 수정",
        "deprecated": false,
        "description": "프리랜서 프로필 정보를 변경합니다.\n\n검수자 : 영은, 경남",
        "tags": [],
        "parameters": [
          {
            "name": "accessToken",
            "in": "cookie",
            "description": "",
            "required": false,
            "example": "<JWT_ACCESS_TOKEN>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "birthDate": {
                    "type": "string",
                    "format": "date",
                    "description": "생일",
                    "nullable": true
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "MALE",
                      "FEMALE"
                    ],
                    "description": "성별",
                    "nullable": true
                  },
                  "businessSector": {
                    "type": "string",
                    "maxLength": 100,
                    "description": "업종",
                    "nullable": true
                  },
                  "businessSectorYears": {
                    "type": "integer",
                    "minimum": -9223372036854776000,
                    "maximum": 9223372036854776000,
                    "description": "업력",
                    "nullable": true
                  },
                  "businessRegistrationNumber": {
                    "type": "string",
                    "description": "사업자등록번호"
                  }
                },
                "required": [
                  "businessRegistrationNumber"
                ]
              },
              "example": {
                "birthDate": "1987-08-05",
                "gender": "MALE",
                "businessSector": "청소",
                "businessSectorYears": 4,
                "businessRegistrationNumber": "69077934708"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "유저 정보 조회",
        "deprecated": false,
        "description": "프리랜서의 정보를 조회합니다. ( 마이페이지 조회용 )\n\n검수자 : 영은, 경남",
        "operationId": "userId",
        "tags": [
          "user id"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "cookie",
            "description": "",
            "required": false,
            "example": "<JWT_ACCESS_TOKEN>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "유저 PK",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "maxLength": 50,
                      "description": "이름"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "FREELANCER",
                        "CLIENT",
                        "ADMIN"
                      ],
                      "default": "FREELANCER",
                      "description": "유저 role\n",
                      "nullable": true
                    },
                    "phone": {
                      "type": "string",
                      "maxLength": 20,
                      "description": "핸드폰번호 11자리",
                      "nullable": true
                    },
                    "birthDate": {
                      "type": "string",
                      "format": "date",
                      "description": "생년월일",
                      "nullable": true
                    },
                    "gender": {
                      "type": "string",
                      "enum": [
                        "MALE",
                        "FEMALE"
                      ],
                      "description": "성별",
                      "nullable": true
                    },
                    "businessSector": {
                      "type": "string",
                      "maxLength": 100,
                      "description": "업종",
                      "nullable": true
                    },
                    "businessSectorYears": {
                      "type": "integer",
                      "minimum": -9223372036854776000,
                      "maximum": 9223372036854776000,
                      "description": "업력",
                      "nullable": true
                    },
                    "businessRegistrationNumber": {
                      "type": "string",
                      "maxLength": 20,
                      "description": "사업자등록번호\n",
                      "nullable": true
                    }
                  },
                  "required": [
                    "name"
                  ]
                },
                "example": {
                  "name": "이영은",
                  "role": "FREELANCER",
                  "phone": "01023970938",
                  "birthDate": "1998-11-23",
                  "gender": "FEMALE",
                  "businessSector": "디자인",
                  "businessSectorYears": 8,
                  "businessRegistrationNumber": "9325863715"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/contracts/extract": {
      "post": {
        "summary": "계약서 추출",
        "deprecated": false,
        "description": "계약서 메타 데이터 추출 ( OCR + LLM )\nOCR로 JSON 데이터 추출 -> Gemini 로 매핑 결과 추출\n\n검수자 : 영은, 윤미",
        "tags": [],
        "parameters": [
          {
            "name": "accessToken",
            "in": "cookie",
            "description": "",
            "required": false,
            "example": "<JWT_ACCESS_TOKEN>",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string",
                    "description": "pdf,png,jpg 파일\n",
                    "example": ""
                  }
                }
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contractInfo": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "string"
                        },
                        "startDate": {
                          "type": "string",
                          "format": "date"
                        },
                        "endDate": {
                          "type": "string",
                          "format": "date"
                        }
                      },
                      "required": [
                        "title",
                        "startDate",
                        "amount",
                        "endDate"
                      ]
                    },
                    "clientInfo": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "name",
                        "phone"
                      ]
                    },
                    "freelancerInfo": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "account": {
                          "type": "string"
                        },
                        "bank": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "name",
                        "phone",
                        "account",
                        "bank"
                      ]
                    }
                  },
                  "required": [
                    "contractInfo",
                    "clientInfo",
                    "freelancerInfo"
                  ]
                },
                "example": {
                  "contractInfo": {
                    "title": "ㅇㅇ 프로젝트",
                    "amount": "150000",
                    "startDate": "2026-03-08",
                    "endDate": "2025-07-24"
                  },
                  "clientInfo": {
                    "name": "김윤미",
                    "phone": "010-1234-5678"
                  },
                  "freelancerInfo": {
                    "name": "이영은",
                    "phone": "010-8765-4321",
                    "account": "123515651",
                    "bank": "국민은행"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/contracts": {
      "post": {
        "summary": "계약서 등록",
        "deprecated": false,
        "description": "계약서 신규 등록 ( PDF 업로드 포함 )\n\n검수자 : 영은, 윤서, 윤미",
        "tags": [],
        "parameters": [
          {
            "name": "accessToken",
            "in": "cookie",
            "description": "",
            "required": false,
            "example": "<JWT_ACCESS_TOKEN>",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "description": "계약서 업로드인지, 계약서 생성 건인지",
                    "enum": [
                      "CREATED",
                      "UPLOAD"
                    ]
                  },
                  "contractInfo": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "amount": {
                        "type": "number"
                      },
                      "startDate": {
                        "type": "string",
                        "format": "date"
                      },
                      "endDate": {
                        "type": "string",
                        "format": "date"
                      }
                    },
                    "required": [
                      "title",
                      "startDate",
                      "amount",
                      "endDate"
                    ]
                  },
                  "clientInfo": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "phone": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "name",
                      "phone"
                    ]
                  },
                  "freelancerInfo": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "account": {
                        "type": "string"
                      },
                      "bank": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "name",
                      "phone",
                      "account",
                      "bank"
                    ]
                  }
                },
                "required": [
                  "type",
                  "contractInfo",
                  "clientInfo",
                  "freelancerInfo"
                ]
              },
              "example": {
                "type": "UPDATE",
                "contractInfo": {
                  "title": "ㅇㅇ 프로젝트",
                  "amount": 150000,
                  "startDate": "2025-12-05",
                  "endDate": "2025-03-13"
                },
                "clientInfo": {
                  "name": "이영은",
                  "phone": "010-2815-1244"
                },
                "freelancerInfo": {
                  "name": "김윤서",
                  "phone": "010-3923-4054",
                  "account": "135165135845",
                  "bank": "국민은행"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactionId": {
                      "type": "integer",
                      "description": "거래 id"
                    }
                  },
                  "required": [
                    "transactionId"
                  ]
                },
                "example": {
                  "transactionId": 14
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/contracts/{contractId}/signatures": {
      "post": {
        "summary": "서명 등록",
        "deprecated": false,
        "description": "프리랜서/의뢰인 서명 저장\n\n검수자 : 영은, 수연, 윤서",
        "tags": [],
        "parameters": [
          {
            "name": "accessToken",
            "in": "cookie",
            "description": "",
            "required": false,
            "example": "<JWT_ACCESS_TOKEN>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contractId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "signatureUrl": {
                    "type": "string",
                    "description": "클라이언트 서명 url"
                  }
                },
                "required": [
                  "signatureUrl"
                ]
              },
              "example": {
                "signatureUrl": "https://gifted-colon.org/"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "nullable": true
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}